bug squash disabled dragging carousel and cleaned up logic on row
the draggable feature on the carousel does not track the index on the selection pips or set the current slide as the active slide.

I cleaned up the logic on the collapsible list row so that there are no longer phantom nested rows fading in. I had added some logic in an odd spot, so now I cleaned it up and made the logic explicit.